requirements.txt
streamlit==1.9.0
numpy==1.21.2
Pillow==8.3.2
pandas==1.3.3
requests==2.26.0
tensorflow==2.6.0